requirements.txt
dash==2.14.1
dash_bootstrap_components==1.5.0
dash_pivottable==0.0.2
numpy==1.24.4
openai==1.3.6
pandas==1.5.3
plotly==5.15.0
python-dotenv==1.0.0
openpyxl
xlrd
